Major rewrite started
Now uses React to make the UI components. No more Mustache. Render is
much faster.
There is still works to do to display more Card information and to be
able to select what is/isn't displayed and add search. I have to work on
the CSS to make printing prettier and have a look that more ressemble
Trello.
Now uses bower for the newer dependencies.